Fix compile issue
