+++ client/internal/wgproxy/ebpf/proxy_interface.go
@@ -1,0 +1,5 @@
+package ebpf
+
+type Proxy interface {
+	Free() error
+}
